Refactor: moved controllers related to CRUD into the crud package.
