+++ src/main/java/code/rice/bowl/spaghetti/controller/crud/BadgeController.java
@@ -1,4 +1,4 @@
-package code.rice.bowl.spaghetti.controller;
+package code.rice.bowl.spaghetti.controller.crud;
 
 import code.rice.bowl.spaghetti.dto.badge.BadgeRequest;
 import code.rice.bowl.spaghetti.dto.badge.BadgeResponse;
