+++ src/main/java/code/rice/bowl/spaghetti/controller/crud/TopicController.java
@@ -1,4 +1,4 @@
-package code.rice.bowl.spaghetti.controller;
+package code.rice.bowl.spaghetti.controller.crud;
 
 import code.rice.bowl.spaghetti.dto.topic.TopicRequest;
 import code.rice.bowl.spaghetti.dto.topic.TopicResponse;
