+++ src/main/java/code/rice/bowl/spaghetti/controller/crud/UserTopicController.java
@@ -1,4 +1,4 @@
-package code.rice.bowl.spaghetti.controller;
+package code.rice.bowl.spaghetti.controller.crud;
 
 import code.rice.bowl.spaghetti.dto.usertopic.UserTopicResponse;
 import code.rice.bowl.spaghetti.service.UserTopicService;
